Moves functions into their own files
